Refactor the distillation by moving functions to the shared logic
This is important for the upcoming distillation-based MCP tools.

To verify, as usual just run it with the example, e.g.:

```
./getgather.py run https://goodreads.com/review
```

and then supply the credentials until the loop terminates with the HTML
version of the bookshelf.
